requirements.txt
flask>=3.0.0,<4.0.0
functions>=0.7.0,<0.8.0
openai>=1.2.3,<1.3.0
packaging>=23.2,<24.0
requests>=2.31.0,<2.32.0
python-dotenv==1.0.0